Add the salt-encrypt algorithm defined by the RADIUS WG back in the day.
This is based on the original code from Gabriel Blanchard, but with significant
changes to make it more robust. I've also moved the implemention into the PHP
layer, rather than the radlib layer, since it depends on PHP functionality to
run and hence needs access to TSRM. The following commit will actually wire
this up to the various attribute setters.

I've also added a radius_salt_encrypt_attr() function that basically exposes
the internal function to user space. I'm not sure it's overly useful in
practice, but it does make testing a tonne easier.