Refactor tagging into an options parameter in radlib.
When I implemented tagging, I should really have taken a step back and figured
out the best way to do it rather than just hacking it in. There wasn't really a
need to implement parallel radlib functions for this: it kind of works if tags
are the only attribute option, but with salting coming in as well, it just
didn't make sense.

On the bright side, this actually results in a net loss of code with the test
suite still passing just fine. That's the sign of a good commit. I hope.