Move salting back into radlib.
My main motivation for putting it in radius.c was to avoid TSRMLS_FETCH(),
which is kind of ugly and slow, but realistically it belongs here. I was also
concerned about the dependency on PHP code, but the MPPE demangling code
already depended on PHP's internal MD5 implementation, so we're no worse off
there.